move eval_path to lib_util and rename it to find_files
